feature_comments v1

+++ ewm-service/src/main/java/ru/practicum/user/service/UserServiceImpl.java
@@ -44,7 +44,11 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(Long id) {
-        User user = userRepository.findById(id).orElseThrow(() -> new NotFoundException("Not found"));
-        userRepository.delete(user);
+        if (userRepository.existsById(id)) {
+            User user = userRepository.findById(id).orElseThrow(() -> new NotFoundException("Not found"));
+            userRepository.delete(user);
+        } else {
+            throw new NotFoundException("Not found");
+        }
     }
 }
